Set stacklevel for deprecation warnings.
This allow the warning to be triggered in the right frame (the caller).
In particular when dependees run pytest with

    filterwarnings = [
        "error:::my_package"]
    ....

This make using a deprecated option an error in the test suite, and make
it easier to find as even when not in the test suite the file and line
with the error will be properly highlighted. Like:

DeprecationWarning: TiffWriter: the 'compress' parameter is deprecated
  imwrite('foo.tif', np.array(0), compress=0)

Instead of showing the `warning.warn` line.

In interactive session, IPython shows deprecation warnings in main
namespace by default, meaning that with this change, someone using
imwrite(compress=..) in IPython will also see the error message; it
looks like with master they do not.

In general I believe this would make the migration from compress to
compression and other deprecated keyword faster.

I also like to add the version since it's deprecated to know how pressing
migrating is, or how likely this is something I can always do or if I
might have to worry about backward compatibility