Allow Adding "FillOrder" Tag to File
At the moment it's not allowed to add the tag 266 ("FillOrder") to the tiff file output. When you add a related tuple (e.g. `(TIFF.TAGS.get("FillOrder"), TIFF.DATATYPES.SHORT, 1, TIFF.FILLORDER.MSB2LSB, True)` to the argument `extratags` of the `imwrite` function, this entry will ignores because 266 is part of `TIFF.TAG_FILTERED` (see also line 2978).

I tested it locally: I removed 266 from that list and it works for me. Is there a reason because it's part of `TIFF.TAG_FILTERED`? Otherwise please remove. Thx! :)