Read PlaneProperty from STK files.
The format of PlaneProperty (UIC1 tag 49) is undocumented but appears
to match the one of ImagePropertyEx (UIC1 tag 48... which was previously
implemented but unused).  Note that it appears that tag names can be
repeated (in particular, metadata for multiple planes just get all
concatenated into PlaneProperty entries), hence the choice to return a
list of entries rather than a dict.  I also chose to return strings as
unicode strings rather than bytes, for practicality.